+++ packages/content-blocks/src/Controller/ClipboardController.php
@@ -37,7 +37,6 @@ use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
  *
  * @internal the routes are the contract, not this class
  */
-#[Route('/_content-blocks')]
 final class ClipboardController
 {
     use CsrfProtectedTrait;